requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.4.0
torch>=1.13.0
Pillow>=9.0.0
ipython>=7.0.0  # Optional, for Jupyter Notebook support
jupyter>=1.0.0  # Optional, for running .ipynb notebooks